Modularize decorator "events" and move webclient specific logic out.
After some discussions with Will and Carlos we decided to extensively
modularize the decorator's "events" or "actions". This allows us to be
fairly liberal about where and how the decorator is used. Especially
since the semantics of the webgateway, webadmin, webclient, etc. Django
applications is so different.

Decorator class extension is now largely required to perform more
advanced interactions as part of the view method preparation process.


Former-commit-id: 43fad668b8d5ef7648f54af12b90066e86ea168e