Just use server_id and allow useragent to be changed.
Server model objects cannot be Pickled and when Django encounters a
Pickle.loads() exception it creates a completely new session dictionary.
In order to combat this we will not be storing the Server as an instance
variable when server_id is set.  Furthermore to aid configurability the
user agent is now passed along on all calls to create_connection() or
join_connection().


Former-commit-id: 24319f2a1a41808a83d2a02c4f2e304b1953f166