Logic to allow the configuration of additional applications.
With the moves outlined in #8022 it is critical that additional
applications are able to be added and removed at will during development
without actually modifying omeroweb.settings or omeroweb.urls.  This
commit allows this through the addition of an omero.web.apps
configuration option which loads that application, by name, based on its
presence under the omeroweb directory.  As an example:

    bin/omero config set '["webpublic"]'

This application then has its templates loaded from
omeroweb/webpublic/templates and its URLs inserted based on the
existence of omeroweb/webpublic/urls.py mounted at /webpublic.


Former-commit-id: 33bf50692661c89a96866ec19fbc4b4ad52fea02