Removing 'url' from context of all pages.
I think {{ url }} was only used before for action forms where we needed to redirect
back to the original page in the right-hand panel iframe.
Can't see that this is used anywhere in the templates now. Used in a couple
of places in view methods, forms etc (not removed from these)


Former-commit-id: 582b17c551f26eadee9a42628f695a5ee76d091c