Add webstart configuration (Fix #9478)
Initially, the webstart_jar configuration option was
added in order to allow proper ordering of the insight
jars. However, it made sense to configure all the
hard-coded properties at the same time.

This commit also includes two currently unused properties:

  jnlp.omero.host and jnlp.omero.port

which can be used to select an initial server in insight.


Former-commit-id: 9f4e6e065dd954e9100ad201381ee203a7c03d56