ome.viewportImage.js passes jsHint validation.
As an example of how strict jsHint is, these are the necessary edits to the above file
to get it to validate (given the jshint config file from the previous commit.
Most common changes are adding radial to parseInt(x, 10) and missing semi-colons.
One BUG found was missing {} after an if statement (Carlos mentioned this already?).


Former-commit-id: 85ecef1849bb3153b968e21ada2e3546b8a9bf23