container.py 'acquisition' doesn't load plate too. See PR 491
This fixes the display of Annotations for PlateAcquisitions. Since container.py annotationList() uses the
first object that is not None (E.g. self.plate, self.acquisition) etc, if we load the plate when dealing
with acquisition, then we list the annotations on plate instead of acquisition. So, better to only
load the acquisition (and load the plate if needed, E.g. for getNumberOfFields() ).


Former-commit-id: 7b52643206642e948a1bebc88b6ccc8435ca1a19