Making fixture params immutable
There was a bug/feature in pytest that forced fixture parameters to be immutable, and although this is fixed in current versions I still found this issue in my version, which means others might get it too. Since it isn't that big of a change to play nice with older versions it might be worth applying this change.

https://bitbucket.org/hpk42/pytest/issue/290/

Former-commit-id: 8fffad3ca8621ae5a27d75d6f423e90a6021cb79