Set Dataset permissions rather than Project permissions
It is possible that the permissions of the Project are different than
the permissions of the Dataset.  In order to not have "permission
denied" error dialogs pop up when this situation is run across this
change needs to be made.  It follows the changes already made to the
permission loading semantics from gh-2844.


Former-commit-id: 20ef2f124f80c4e77f5d8b9a35415d1bfdaf52d1