Move OME.handleChgrp() and OME.handleDelete() from containers.html
These functions had been updated to work with new jsTree in containers.html but had in the meantime
been ported to other js files in the main branch (and not resolved on merge). This commit
removes the updated functions from containers.html and moves the changes to the other js files.
Also improves finding of fileset images in jsTree.


Former-commit-id: 2ced732da6f8c6e246cb98513c234e8b0cf033b7