Do not encapsulate null connection behavior in try/finally block
If connectoy.create_gateway fails to create a connection, the contract is to
return None. In the current implementation, the final connection.close() statement will fail.
Moving the nullity tests out of the try/finally blocks should restore the
previous behavior e.g. when performing a version check while preserving the
termination of connections when such a connection exist.


Former-commit-id: 716b93775de55ff591a7ffdc79894de154ef1a69