Treat annotation datetimes as UTC for consistency with metadata
metadata_general.html displays datetimes "as is", i.e. it assumes the datetime returned by OMERO.server is exactly what should be displayed. In contrast OME.formatDate assumes the date returned by OMERO.server is UTC and should be converted for local display. By using javascript Date.getUTC* this ensure the date is displayed unchanged, so as to match metadata_general.html

In practice this will often not be the expected time, but in the absence of a known timezone in OMERO.server this should ensure OMERO.web datetimes are consistently wrong.


Former-commit-id: 961e4ce0204f1d1c557ef891805283bceded4c1e